modified data type

+++ handle_pointers.c
@@ -16,7 +16,7 @@ int handle_pointer(__attribute__((unused)) const format_specifier * spec,
 	int characters_printed;
 	size_t initial_length = buffer->length;
 	unsigned long int pointer = (unsigned long int) va_arg(
-		args, unsigned long int *);
+		args, unsigned int *);
 
 	utob(pointer, hex_str, HEX);
 	append_string(buffer, "0x");
